wip: refactor table_row_selector to referentiel_de_polynesie
