+++ app/components/dossiers/champs_rows_show_component/champs_rows_show_component.html.haml
@@ -56,8 +56,8 @@
             = render partial: "shared/champs/te_fenua/show", locals: { champ: champ, profile: @profile }
           - when TypeDeChamp.type_champs.fetch(:numero_dn)
             = render partial: "shared/champs/numero_dn/show", locals: { champ: champ, profile: @profile }
-          - when TypeDeChamp.type_champs.fetch(:table_row_selector)
-            = render partial: "shared/champs/table_row_selector/show", locals: { champ: champ, profile: @profile }
+          - when TypeDeChamp.type_champs.fetch(:referentiel_de_polynesie)
+            = render partial: "shared/champs/referentiel_de_polynesie/show", locals: { champ: champ, profile: @profile }
           - when TypeDeChamp.type_champs.fetch(:cojo)
             = render partial: "shared/champs/cojo/show", locals: { champ: champ, profile: @profile }
           - when TypeDeChamp.type_champs.fetch(:date)
