+++ app/components/editable_champ/table_row_selector_component.rb
@@ -1,12 +1,12 @@
-class EditableChamp::TableRowSelectorComponent < EditableChamp::ComboSearchComponent
-  def dsfr_input_classname
-    'fr-input'
-  end
+# class EditableChamp::TableRowSelectorComponent < EditableChamp::ComboSearchComponent
+#   def dsfr_input_classname
+#     'fr-input'
+#   end
 
-  def react_input_opts
-    opts = input_opts(id: @champ.input_id, required: @champ.required?, aria: { describedby: @champ.describedby_id }, scopeExtra: @champ.table_id)
-    opts[:className] = "#{opts.delete(:class)} fr-mt-1w"
+#   def react_input_opts
+#     opts = input_opts(id: @champ.input_id, required: @champ.required?, aria: { describedby: @champ.describedby_id }, scopeExtra: @champ.table_id)
+#     opts[:className] = "#{opts.delete(:class)} fr-mt-1w"
 
-    opts
-  end
-end
+#     opts
+#   end
+# end
