+++ app/components/editable_champ/table_row_selector_component/table_row_selector_component.html.haml
@@ -1,7 +1,7 @@
-- render_parent
+-# - render_parent
 
-= @form.hidden_field :value
-= @form.hidden_field :external_id
-= react_component("ComboTableRowSelectorSearch",
-  **react_input_opts,
-  **react_combo_props)
+-# = @form.hidden_field :value
+-# = @form.hidden_field :external_id
+-# = react_component("ComboTableRowSelectorSearch",
+-#   **react_input_opts,
+-#   **react_combo_props)
