+++ app/controllers/champs/table_row_selector_controller.rb
@@ -1,20 +1,20 @@
-# frozen_string_literal: true
+# # frozen_string_literal: true
 
-class Champs::TableRowSelectorController < ApplicationController
-  before_action :authenticate_logged_user!
+# class Champs::TableRowSelectorController < ApplicationController
+#   before_action :authenticate_logged_user!
 
-  def search
-    @params = search_params
-    if bad_parameters
-      render json: [], status: 400
-    else
-      render json: TableRowSelector::API.search(@params[:domain], @params[:term])
-    end
-  end
+#   def search
+#     @params = search_params
+#     if bad_parameters
+#       render json: [], status: 400
+#     else
+#       render json: TableRowSelector::API.search(@params[:domain], @params[:term])
+#     end
+#   end
 
-  def bad_parameters
-    @params[:domain].blank? || @params[:domain].to_i == 0 || @params[:term].blank?
-  end
+#   def bad_parameters
+#     @params[:domain].blank? || @params[:domain].to_i == 0 || @params[:term].blank?
+#   end
 
-  def search_params = params.permit(:domain, :term)
-end
+#   def search_params = params.permit(:domain, :term)
+# end
