+++ app/models/champs/referentiel_de_polynesie_champ.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+class Champs::ReferentielDePolynesieChamp < Champs::TextChamp
+  def fetch_external_data?
+    true
+  end
+
+  def fetch_external_data
+    # APIEducation::AnnuaireEducationAdapter.new(external_id).to_params
+    ReferentielDePolynesie::API.fetch_row(external_id)
+  end
+
+  def update_with_external_data!(data:)
+    update!(data: data) if data&.is_a?(Hash)
+  end
+end
