+++ app/models/logic/champ_value.rb
@@ -1,5 +1,5 @@
 class Logic::ChampValue < Logic::Term
-  INSTANCE_MANAGED_TYPE_DE_CHAMP = [:table_row_selector]
+  INSTANCE_MANAGED_TYPE_DE_CHAMP = [:referentiel_de_polynesie]
 
   MANAGED_TYPE_DE_CHAMP = TypeDeChamp.type_champs.slice(
     *INSTANCE_MANAGED_TYPE_DE_CHAMP,
@@ -71,7 +71,9 @@ class Logic::ChampValue < Logic::Term
         code_departement: targeted_champ.code_departement,
         code_region: targeted_champ.code_region
       }
-    when "Champs::TableRowSelectorChamp"
+    # when "Champs::TableRowSelectorChamp"
+    #   targeted_champ.value
+    when "Champs::ReferentielDePolynesieChamp"
       targeted_champ.value
     when "Champs::CommuneDePolynesieChamp", "Champs::CodePostalDePolynesieChamp"
       {
@@ -104,7 +106,7 @@ class Logic::ChampValue < Logic::Term
       CHAMP_VALUE_TYPE.fetch(:departement_enum)
     when MANAGED_TYPE_DE_CHAMP.fetch(:multiple_drop_down_list)
       CHAMP_VALUE_TYPE.fetch(:enums)
-    when MANAGED_TYPE_DE_CHAMP.fetch(:table_row_selector)
+    when MANAGED_TYPE_DE_CHAMP.fetch(:referentiel_de_polynesie)
       CHAMP_VALUE_TYPE.fetch(:enum)
     else
       CHAMP_VALUE_TYPE.fetch(:unmanaged)
@@ -143,7 +145,7 @@ class Logic::ChampValue < Logic::Term
       APIGeoService.departements.map { ["#{_1[:code]} – #{_1[:name]}", _1[:code]] }
     elsif operator_name.in?([Logic::InArchipelOperator.name, Logic::NotInArchipelOperator.name]) || tdc.type_champ.in?([MANAGED_TYPE_DE_CHAMP.fetch(:commune_de_polynesie), MANAGED_TYPE_DE_CHAMP.fetch(:code_postal_de_polynesie)])
       APIGeo::API.archipels_de_polynesie.map { [_1, _1] }
-    elsif tdc.type_champ == MANAGED_TYPE_DE_CHAMP.fetch(:table_row_selector)
+    elsif tdc.type_champ == MANAGED_TYPE_DE_CHAMP.fetch(:referentiel_de_polynesie)
       [['Autre', 'Autre']]
     else
       tdc.drop_down_list_enabled_non_empty_options(other: true).map { _1.is_a?(Array) ? _1 : [_1, _1] }
