+++ app/models/types_de_champ/referentiel_de_polynesie_type_de_champ.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+class TypesDeChamp::ReferentielDePolynesieTypeDeChamp < TypesDeChamp::TypeDeChampBase
+end
