--- app/models/types_de_champ/table_row_selector_type_de_champ.rb
@@ -1,4 +1,0 @@
-# frozen_string_literal: true
-
-class TypesDeChamp::TableRowSelectorTypeDeChamp < TypesDeChamp::TypeDeChampBase
-end
