+++ config/locales/shared.en.yml
@@ -17,7 +17,7 @@ en:
           refuse: "Expires on %{date} (%{duree_conservation_totale} months after this file was processed)"
           sans_suite: "Expires on %{date} (%{duree_conservation_totale} months after this file was processed)"
     champs:
-      table_row_selector:
+      referentiel_de_polynesie:
         show:
           toggle_champ: "Show/hide fields of « %{champ} »"
       cnaf:
