+++ config/routes.rb
@@ -204,7 +204,7 @@ Rails.application.routes.draw do
     post ':dossier_id/:stable_id/repetition', to: 'repetition#add', as: :repetition
     delete ':dossier_id/:stable_id/repetition', to: 'repetition#remove'
 
-    get 'table_row_selector/search', to: 'table_row_selector#search'
+    get 'referentiel_de_polynesie/search', to: 'referentiel_de_polynesie#search'
 
     get ':dossier_id/:stable_id/siret', to: 'siret#show', as: :siret
     get ':dossier_id/:stable_id/rna', to: 'rna#show', as: :rna
