+++ spec/factories/type_de_champ.rb
@@ -199,6 +199,9 @@ FactoryBot.define do
     factory :type_de_champ_visa do
       type_champ { TypeDeChamp.type_champs.fetch(:visa) }
     end
+    factory :type_de_champ_referentiel_de_polynesie do
+      type_champ { TypeDeChamp.type_champs.fetch(:referentiel_de_polynesie) }
+    end
     factory :type_de_champ_epci do
       type_champ { TypeDeChamp.type_champs.fetch(:epci) }
     end
