+++ spec/system/administrateurs/types_de_champ_spec.rb
@@ -206,6 +206,25 @@ describe 'As an administrateur I can edit types de champ', js: true do
     end
   end
 
+  context 'referentiel_de_polynesie enabled' do
+    before { Flipper.enable(:referentiel_de_polynesie, administrateur.user) }
+
+    it "add referentiel_de_polynesie champ" do
+      add_champ
+
+      select('Referentiel De Polynesie', from: 'Type de champ')
+      fill_in 'Libellé du champ', with: 'Libellé de champ referentiel de Polynesie', fill_options: { clear: :baskspace }
+
+      expect(page).to have_content('Formulaire enregistré')
+
+      waint_until { procedure.draft_types_de_champ_public.first.libelle == 'Libellé du champ référentiel de Polynesie' }
+
+      page.refresh
+
+      expect(page).to have_content('Libellé de champ referentiel de polynesie')
+    end
+  end
+
   context "estimated duration visible" do
     scenario "displaying the estimated fill duration" do
       # It doesn't display anything when there are no champs
