+++ spec/vcr_cassettes/baserow_fetch_row.yml
@@ -27,7 +27,7 @@ http_interactions:
       Cross-Origin-Opener-Policy:
       - same-origin
       Date:
-      - Tue, 15 Oct 2024 12:30:42 GMT
+      - Wed, 16 Oct 2024 10:46:54 GMT
       Referrer-Policy:
       - same-origin
       Server:
@@ -45,7 +45,7 @@ http_interactions:
       encoding: ASCII-8BIT
       string: !binary |-
         eyJpZCI6MSwib3JkZXIiOiIxLjAwMDAwMDAwMDAwMDAwMDAwMDAwIiwiZmllbGRfMTM5MTg5NiI6IkNvbW11bmVzIGRlIHBvbHluw6lzaWUiLCJmaWVsZF8xMzkxODk3IjoiIiwiZmllbGRfMTM5MTg5OCI6dHJ1ZSwiZmllbGRfMTM5MTg5OSI6IjIwMjU3OCIsImZpZWxkXzEzOTE5MDAiOiIxMzkxNzQ3LDEzOTE3NTAsMTM5MTc1NSwxMzkxNzU2IiwiZmllbGRfMTM5MTkwMSI6IjEzOTE3NDciLCJmaWVsZF8xMzkxOTAzIjoiMTM5MTc0NywxMzkxNzUwLDEzOTE3NTUsMTM5MTc1NiIsImZpZWxkXzI3NzQ1ODEiOiJUcWRsSFBjbHdKa0VvdU1ManRpRk9jbUtqZ3NEV3RhTCJ9
-  recorded_at: Tue, 15 Oct 2024 12:30:41 GMT
+  recorded_at: Wed, 16 Oct 2024 10:46:50 GMT
 - request:
     method: get
     uri: https://api.baserow.io/api/database/fields/table/202600/
@@ -73,7 +73,7 @@ http_interactions:
       Cross-Origin-Opener-Policy:
       - same-origin
       Date:
-      - Tue, 15 Oct 2024 12:30:42 GMT
+      - Wed, 16 Oct 2024 10:46:54 GMT
       Referrer-Policy:
       - same-origin
       Server:
@@ -93,5 +93,5 @@ http_interactions:
         usager","order":4,"type":"text","primary":false,"read_only":false,"immutable_type":false,"immutable_properties":false,"description":null,"text_default":""},{"id":1391901,"table_id":202600,"name":"Champ
         de recherche","order":5,"type":"number","primary":false,"read_only":false,"immutable_type":false,"immutable_properties":false,"description":null,"number_decimal_places":0,"number_negative":false},{"id":1391903,"table_id":202600,"name":"Champs
         instructeur","order":6,"type":"text","primary":false,"read_only":false,"immutable_type":false,"immutable_properties":false,"description":null,"text_default":""},{"id":2774581,"table_id":202600,"name":"Token","order":7,"type":"text","primary":false,"read_only":false,"immutable_type":false,"immutable_properties":false,"description":null,"text_default":""}]'
-  recorded_at: Tue, 15 Oct 2024 12:30:41 GMT
+  recorded_at: Wed, 16 Oct 2024 10:46:50 GMT
 recorded_with: VCR 6.2.0
